+++ app/components/UI/EthInput/index.js
@@ -19,12 +19,13 @@ import { strings } from '../../../../locales/i18n';
 import TokenImage from '../TokenImage';
 import MaterialIcon from 'react-native-vector-icons/MaterialIcons';
 import ElevatedView from 'react-native-elevated-view';
-import CollectibleImage from '../CollectibleImage';
+import CollectibleMedia from '../CollectibleMedia';
 import SelectableAsset from './SelectableAsset';
 import { getTicker, getNormalizedTxState } from '../../../util/transactions';
 import FontAwesome from 'react-native-vector-icons/FontAwesome';
 import Device from '../../../util/Device';
 import NetworkMainAssetLogo from '../NetworkMainAssetLogo';
+import { toLowerCaseCompare } from '../../../util/general';
 
 const styles = StyleSheet.create({
 	root: {
@@ -282,8 +283,8 @@ class EthInput extends PureComponent {
 				});
 				break;
 			case 'CONTRACT_COLLECTIBLE_TRANSACTION': {
-				const collectiblesToShow = collectibles.filter(
-					collectible => collectible.address.toLowerCase() === transaction.selectedAsset.address.toLowerCase()
+				const collectiblesToShow = collectibles.filter(collectible =>
+					toLowerCaseCompare(collectible.address, transaction.selectedAsset.address)
 				);
 				this.setState({
 					assets: collectiblesToShow
@@ -344,7 +345,7 @@ class EthInput extends PureComponent {
 				const subTitle =
 					strings('collectible.collectible_token_id') + strings('unit.colon') + ' ' + asset.tokenId;
 				const icon = (
-					<CollectibleImage collectible={asset} containerStyle={styles.logo} iconStyle={styles.logo} />
+					<CollectibleMedia small collectible={asset} containerStyle={styles.logo} iconStyle={styles.logo} />
 				);
 				return { title, subTitle, icon };
 			}
@@ -629,7 +630,8 @@ class EthInput extends PureComponent {
 							title={selectedAsset.name}
 							subTitle={`${strings('unit.token_id')}${selectedAsset.tokenId}`}
 							icon={
-								<CollectibleImage
+								<CollectibleMedia
+									small
 									collectible={selectedAsset}
 									containerStyle={styles.logo}
 									iconStyle={styles.logo}