+++ app/components/UI/TransactionElement/utils.js
@@ -25,6 +25,7 @@ import contractMap from '@metamask/contract-metadata';
 import { toChecksumAddress } from 'ethereumjs-util';
 import { swapsUtils } from '@metamask/swaps-controller';
 import { isSwapsNativeAsset } from '../Swaps/utils';
+import { toLowerCaseCompare } from '../../../util/general';
 
 const { getSwapsContractAddress } = swapsUtils;
 
@@ -138,9 +139,7 @@ function getCollectibleTransfer(args) {
 	let actionKey;
 	const [, tokenId] = decodeTransferData('transfer', data);
 	const ticker = getTicker(args.ticker);
-	const collectible = collectibleContracts.find(
-		collectible => collectible.address.toLowerCase() === to.toLowerCase()
-	);
+	const collectible = collectibleContracts.find(collectible => toLowerCaseCompare(collectible.address, to));
 	if (collectible) {
 		actionKey = `${strings('transactions.sent')} ${collectible.name}`;
 	} else {
@@ -336,17 +335,15 @@ function decodeTransferFromTx(args) {
 		selectedAddress
 	} = args;
 	const [addressFrom, addressTo, tokenId] = decodeTransferData('transferFrom', data);
-	const collectible = collectibleContracts.find(
-		collectible => collectible.address.toLowerCase() === to.toLowerCase()
-	);
+	const collectible = collectibleContracts.find(collectible => toLowerCaseCompare(collectible.address, to));
 	let actionKey = args.actionKey;
 	if (collectible) {
 		actionKey = `${strings('transactions.sent')} ${collectible.name}`;
 	}
 
 	const totalGas = calculateTotalGas(gas, gasPrice);
-	const renderCollectible = collectible
-		? `${strings('unit.token_id')}${tokenId} ${collectible.symbol}`
+	const renderCollectible = collectible?.symbol
+		? `${strings('unit.token_id')}${tokenId} ${collectible?.symbol}`
 		: `${strings('unit.token_id')}${tokenId}`;
 
 	const renderFrom = renderFullAddress(addressFrom);