+++ app/components/Views/SendFlow/Amount/index.js
@@ -27,7 +27,6 @@ import {
 	renderFromWei,
 	weiToFiat,
 	fromWei,
-	fromTokenMinimalUnit,
 	toWei,
 	isDecimal,
 	toTokenMinimalUnit,
@@ -36,14 +35,15 @@ import {
 	weiToFiatNumber,
 	balanceToFiatNumber,
 	getCurrencySymbol,
-	handleWeiNumber
+	handleWeiNumber,
+	fromTokenMinimalUnitString
 } from '../../../../util/number';
 import { getTicker, generateTransferData, getEther } from '../../../../util/transactions';
 import { util } from '@metamask/controllers';
 import ErrorMessage from '../ErrorMessage';
 import { getGasPriceByChainId } from '../../../../util/custom-gas';
 import Engine from '../../../../core/Engine';
-import CollectibleImage from '../../../UI/CollectibleImage';
+import CollectibleMedia from '../../../UI/CollectibleMedia';
 import collectiblesTransferInformation from '../../../../util/collectibles-transfer';
 import { strings } from '../../../../../locales/i18n';
 import Device from '../../../../util/Device';
@@ -53,6 +53,7 @@ import { ANALYTICS_EVENT_OPTS } from '../../../../util/analytics';
 import dismissKeyboard from 'react-native/Libraries/Utilities/dismissKeyboard';
 import NetworkMainAssetLogo from '../../../UI/NetworkMainAssetLogo';
 import { isMainNet } from '../../../../util/networks';
+import { toLowerCaseCompare } from '../../../../util/general';
 
 const { hexToBN, BNToHex } = util;
 
@@ -241,7 +242,7 @@ const styles = StyleSheet.create({
 	errorMessageWrapper: {
 		marginVertical: 16
 	},
-	collectibleImage: {
+	CollectibleMedia: {
 		width: 120,
 		height: 120
 	},
@@ -423,7 +424,7 @@ class Amount extends PureComponent {
 		} = this.props;
 		try {
 			const owner = await AssetsContractController.getOwnerOf(address, tokenId);
-			const isOwner = owner.toLowerCase() === selectedAddress.toLowerCase();
+			const isOwner = toLowerCaseCompare(owner, selectedAddress);
 			if (!isOwner) {
 				return strings('transaction.invalid_collectible_ownership');
 			}
@@ -656,10 +657,16 @@ class Amount extends PureComponent {
 		} else {
 			const exchangeRate = contractExchangeRates[selectedAsset.address];
 			if (internalPrimaryCurrencyIsCrypto || !exchangeRate) {
-				input = fromTokenMinimalUnit(contractBalances[selectedAsset.address], selectedAsset.decimals);
+				input = fromTokenMinimalUnitString(
+					contractBalances[selectedAsset.address]?.toString(10),
+					selectedAsset.decimals
+				);
 			} else {
 				input = `${balanceToFiatNumber(
-					fromTokenMinimalUnit(contractBalances[selectedAsset.address], selectedAsset.decimals),
+					fromTokenMinimalUnitString(
+						contractBalances[selectedAsset.address]?.toString(10),
+						selectedAsset.decimals
+					),
 					conversionRate,
 					exchangeRate
 				)}`;
@@ -827,7 +834,8 @@ class Amount extends PureComponent {
 				onPress={() => this.pickSelectedAsset(collectible)}
 			>
 				<View style={styles.assetElement}>
-					<CollectibleImage
+					<CollectibleMedia
+						small
 						collectible={collectible}
 						iconStyle={styles.tokenImage}
 						containerStyle={styles.tokenImage}
@@ -965,9 +973,10 @@ class Amount extends PureComponent {
 		return (
 			<View style={styles.collectibleInputWrapper}>
 				<View style={styles.collectibleInputImageWrapper}>
-					<CollectibleImage
-						containerStyle={styles.collectibleImage}
-						iconStyle={styles.collectibleImage}
+					<CollectibleMedia
+						small
+						containerStyle={styles.CollectibleMedia}
+						iconStyle={styles.CollectibleMedia}
 						collectible={selectedAsset}
 					/>
 				</View>